Support rrrr/gggg/bbbb escape sequence colors
The rrrr/gggg/bbbb sequence is common in responses to
foreground/background color requests. Formatting responses to these
requests in this manner will aid in compatibility. Since the xterm
specification indicates that the reponse to foreground/background color
requests must be in the format of an escape sequence which can set the
foreground background color, rrrr/gggg/bbbb should be treated as a valid
escape sequence color format universally.